Move relative urls in index.html one level higher
The issue was that /exports is treated as a resource while /exports/ is treated as a directory, so requests were being made for /exports/build/bundle.js, etc., which 404'd. Adding double dots to each relative URL is a simple fix, since attempting to go above the root directory is blocked.