Adjusted verb, description, and level payload for each notification
Notifications are sent when a job is created. when it is rerun, and when it is complete.  When complete, the verb is END, the level is success or warn and the description matches the run's final status - COMPLETE, CANCELED, INCOMPLETE

Also adjusted an endpoint name to mark_all_as_read